added first logic of plus and minus realization

+++ include/my_exception.hpp
@@ -4,14 +4,14 @@
 
 class MyException {
   public:
-    MyException(int32_t code, std::string error);
+    MyException(int64_t code, std::string error);
 
     [[nodiscard]]
-    int32_t get_code() const;
+    int64_t get_code() const;
     [[nodiscard]]
     std::string get_error() const;
 
   private:
-    int32_t m_code;
+    int64_t m_code;
     std::string m_error;
 };
